requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
python-dotenv
tiktoken
unstructured[local-inference]
sentence-transformers
groq
transformers
torch
